hw6/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
finnhub-python==2.4.8
Flask==2.0.2
google-api-core==2.4.0
google-auth==2.5.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
gunicorn==20.0.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
